+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Support for ACDSee caption and categories metadata
+  [#67](https://github.com/gino0631/nextcloud-metadata/issues/67)
+
 ### Fixed
 - PHP 7.4 errors fixed by updating GetID3 to 1.9.20
   [#62](https://github.com/gino0631/nextcloud-metadata/issues/62)